test: update structure
